Move validation functionality to its own file for now while it's still in a fluctuating state.
